Add files via upload
This file is a connection of sensors to our microcontroller. Since we could not find any proteus libraries for ph level, electrical conductivity and water level sensor, we decided to use available sensors to demonstrate the analog to digital conversion of the sensors that we don't have. We will adjust input values for the replaced sensors we used to control for our intended purpose.